Split test file in two: model and manager
